e_table1_summaries, file.remove retry when too fast for OS

+++ R/e_table1_summaries.R
@@ -7,10 +7,10 @@
 #'    Uses \code{moonBook::mytable}; default is CatMethod = 2: Numeric summaries are Median and IQR with kruskal.test, Categorical summaries are proportions with chisq.test with continuity correction.
 #'
 #' @param dat_tab1            Dataset for table.
-#' @param list_var_col_table  Column variable name.
-#' @param list_var_col_names  Column variable label.
-#' @param list_var_row_table  Variable names to loop over.
-#' @param list_var_row_names  Variable labels to label table.
+#' @param list_var_col_names  Column variable name.
+#' @param list_var_col_labels Column variable label.
+#' @param list_var_row_names  Variable names to loop over.
+#' @param list_var_row_labels Variable labels to label table.
 #' @param fn_root             Filename for temporary tables.
 #' @param fn_all              Filename for output table, defaults to \code{paste0(fn_root, "_", "all", ".csv")}.
 #' @param label_width         Width to wrap column and row variable labels, does not affect category labels.
@@ -31,8 +31,8 @@
 #' ## For Rmd code chunk attractive html table, code chunk option ```{r, results='asis'}
 #'
 #' # specify tows to summarize and columns to summarize by
-#' list_var_col_table <- c("cyl", "am")
-#' list_var_row_table <- c("mpg", "cyl", "disp", "hp", "drat", "wt"
+#' list_var_col_names <- c("cyl", "am")
+#' list_var_row_names <- c("mpg", "cyl", "disp", "hp", "drat", "wt"
 #'                       , "qsec", "vs", "am", "gear", "carb")
 #'
 #' # assigning to new dataset since you may want to filter or subset in some way,
@@ -40,27 +40,27 @@
 #' dat_tab1 <-
 #'   dat_mtcars_e %>%
 #'   #dplyr::filter(EVENT_ID == "BL") %>%
-#'   dplyr::select(all_of(c(list_var_col_table, list_var_row_table)))
+#'   dplyr::select(all_of(c(list_var_col_names, list_var_row_names)))
 #'
 #'
 #' # label variables
-#' list_var_col_names <- labelled::var_label(dat_tab1[, list_var_col_table]) %>% unlist()
-#' list_var_row_names <- labelled::var_label(dat_tab1[, list_var_row_table]) %>% unlist()
+#' list_var_col_labels <- labelled::var_label(dat_tab1[, list_var_col_names]) %>% unlist()
+#' list_var_row_labels <- labelled::var_label(dat_tab1[, list_var_row_names]) %>% unlist()
 #'
 #' # This loop creates a table summarized by each variable,
 #' #   but can also specify multiple columns to summarize conditional on multiple columns.
-#' for (i_col in seq_along(list_var_col_table)) {
+#' for (i_col in seq_along(list_var_col_names)) {
 #'
 #'   # filename root for each separate variable (will write, then read to compile into one)
-#'   fn_root <- paste0("tab1_vars_", list_var_col_table[i_col], "_summary")
+#'   fn_root <- paste0("tab1_vars_", list_var_col_names[i_col], "_summary")
 #'
 #'   # run function (additional options are available)
 #'   e_table1_summaries(
 #'       dat_tab1           = dat_tab1
-#'     , list_var_col_table = list_var_col_table[i_col]
 #'     , list_var_col_names = list_var_col_names[i_col]
-#'     , list_var_row_table = list_var_row_table
+#'     , list_var_col_labels = list_var_col_labels[i_col]
 #'     , list_var_row_names = list_var_row_names
+#'     , list_var_row_labels = list_var_row_labels
 #'     , fn_root            = fn_root
 #'     , sw_verbose         = TRUE
 #'     )
@@ -78,10 +78,10 @@
 #' # run function (additional options are available)
 #' e_table1_summaries(
 #'     dat_tab1           = dat_tab1
-#'   , list_var_col_table = NULL
-#'   , list_var_col_names = "total"
-#'   , list_var_row_table = list_var_row_table
+#'   , list_var_col_names = NULL
+#'   , list_var_col_labels = "total"
 #'   , list_var_row_names = list_var_row_names
+#'   , list_var_row_labels = list_var_row_labels
 #'   , fn_root            = fn_root
 #'   , sw_verbose         = TRUE
 #'   )
@@ -95,10 +95,10 @@
 e_table1_summaries <-
   function(
     dat_tab1            = dat_sub
-  , list_var_col_table  = NULL
   , list_var_col_names  = NULL
-  , list_var_row_table  = NULL
+  , list_var_col_labels  = NULL
   , list_var_row_names  = NULL
+  , list_var_row_labels  = NULL
   , fn_root             = "tab1_vars"
   , fn_all              = NULL
   , label_width         = 40
@@ -108,20 +108,20 @@ e_table1_summaries <-
   , moonbook_method     = c(1, 2, 3)[2]
   ) {
 
-  if (is.null(list_var_row_names)) {
-    list_var_row_names <- list_var_row_table
+  if (is.null(list_var_row_labels)) {
+    list_var_row_labels <- list_var_row_names
   }
 
   #ind_list_csv <- NULL
 
   tab1_list <- NULL
-  for (i_var in seq_along(list_var_row_table)) {
+  for (i_var in seq_along(list_var_row_names)) {
     ## i_var = 5
 
-    if(!is.null(list_var_col_table)) {
-      form <- stats::formula(paste0(paste0(list_var_col_table, collapse = " + "), " ~ ", list_var_row_table[i_var]))
+    if(!is.null(list_var_col_names)) {
+      form <- stats::formula(paste0(paste0(list_var_col_names, collapse = " + "), " ~ ", list_var_row_names[i_var]))
     } else {
-      form <- stats::formula(paste0(" ~ ", list_var_row_table[i_var]))
+      form <- stats::formula(paste0(" ~ ", list_var_row_names[i_var]))
     }
     if (sw_verbose) {
       print(form)
@@ -130,14 +130,14 @@ e_table1_summaries <-
 
 
     # if not a factor, then calculate significant digits to use
-    if (is.factor(dat_tab1[,list_var_row_table[i_var]][[1]])) {
+    if (is.factor(dat_tab1[,list_var_row_names[i_var]][[1]])) {
       temp_digits = moonbook_digits[1] %>% as.numeric()
     } else {
 
       # determine digits to use, when < 1.0
 
       # start with minimum positive absolute value, then log10 + 2 to give 3 sig digits
-      abs_val <- abs(dat_tab1[,list_var_row_table[i_var]])
+      abs_val <- abs(dat_tab1[,list_var_row_names[i_var]])
       abs_val <- abs_val[!is.na(abs_val)]
       abs_val <- abs_val[abs_val > 0]
       min_abs_val <- min(abs_val)
@@ -170,18 +170,18 @@ e_table1_summaries <-
 
     #tab1_list[[i_var]]$res
 
-    if(!is.null(list_var_col_table)) {
+    if(!is.null(list_var_col_names)) {
       # Better name for column
       if (i_var == 1) {
         # first row, only, for long column name
-        if (!is.null(list_var_col_names) & !(list_var_col_names == "NULL") & !(list_var_col_names == "character(0)")) {
-          colnames(tab1_list[[i_var]]$res)[1] <- stringr::str_wrap(list_var_col_names, width = label_width)
+        if (!is.null(list_var_col_labels) & !(list_var_col_labels == "NULL") & !(list_var_col_labels == "character(0)")) {
+          colnames(tab1_list[[i_var]]$res)[1] <- stringr::str_wrap(list_var_col_labels, width = label_width)
         }
       }
       # Better name for variable
-      if (!is.na(list_var_row_names[i_var])) {
-        if (!is.null(list_var_row_names[i_var]) & !(list_var_row_names[i_var] == "NULL") & !(list_var_row_names[i_var] == "character(0)")) {
-          tab1_list[[i_var]]$res[1,1] <- stringr::str_wrap(list_var_row_names[i_var], width = label_width)
+      if (!is.na(list_var_row_labels[i_var])) {
+        if (!is.null(list_var_row_labels[i_var]) & !(list_var_row_labels[i_var] == "NULL") & !(list_var_row_labels[i_var] == "character(0)")) {
+          tab1_list[[i_var]]$res[1,1] <- stringr::str_wrap(list_var_row_labels[i_var], width = label_width)
         }
       }
 
@@ -192,14 +192,14 @@ e_table1_summaries <-
       # # Better name for column
       # if (i_var == 1) {
       #   # first row, only, for long column name
-      #   if (!is.null(list_var_col_names) & !(list_var_col_names == "NULL") & !(list_var_col_names == "character(0)")) {
-      #     colnames(tab1_list[[i_var]]$res)[1] <- stringr::str_wrap(list_var_col_names, width = label_width)
+      #   if (!is.null(list_var_col_labels) & !(list_var_col_labels == "NULL") & !(list_var_col_labels == "character(0)")) {
+      #     colnames(tab1_list[[i_var]]$res)[1] <- stringr::str_wrap(list_var_col_labels, width = label_width)
       #   }
       # }
       # Better name for variable
-      if (!is.na(list_var_row_names[i_var])) {
-        if (!is.null(list_var_row_names[i_var]) & !(list_var_row_names[i_var] == "NULL") & !(list_var_row_names[i_var] == "character(0)")) {
-          tab1_list[[i_var]]$name <- stringr::str_wrap(list_var_row_names[i_var], width = label_width)
+      if (!is.na(list_var_row_labels[i_var])) {
+        if (!is.null(list_var_row_labels[i_var]) & !(list_var_row_labels[i_var] == "NULL") & !(list_var_row_labels[i_var] == "character(0)")) {
+          tab1_list[[i_var]]$name <- stringr::str_wrap(list_var_row_labels[i_var], width = label_width)
         }
       }
     }
@@ -215,8 +215,8 @@ e_table1_summaries <-
 
 
   # save each variable in a separate file to merge together below
-  for (i_var in seq_along(list_var_row_table)) {
-    if(!is.null(list_var_col_table)) {
+  for (i_var in seq_along(list_var_row_names)) {
+    if(!is.null(list_var_col_names)) {
 
       # skip if no table created
       if(is.null(tab1_list[[i_var]])) {
@@ -241,8 +241,8 @@ e_table1_summaries <-
 
   # read each variable file and merge together, then delete original files
   tab1_out <- NULL
-  #for (i_var in ind_list_csv) {  #seq_along(list_var_row_table)) {
-  for (i_var in seq_along(list_var_row_table)) {
+  #for (i_var in ind_list_csv) {  #seq_along(list_var_row_names)) {
+  for (i_var in seq_along(list_var_row_names)) {
     fn_in <- paste0(fn_root, "_", i_var, ".csv")
     if (sw_verbose) {
       print(fn_in)
@@ -266,7 +266,18 @@ e_table1_summaries <-
       print(temp_in)
     }
     tab1_out <- c(tab1_out, temp_in)
-    file.remove(fn_in)
+
+    # If couldn't remove file (because too fast for operating system),
+    #   then sleep and try again
+    check_message <-
+      e_message_capture(
+        file.remove(fn_in)
+      )(1)
+    if (stringr::str_detect(string = check_message$logs[[1]]$message, pattern = stringr::fixed("Permission denied")) ) {
+      message("erikmisc::e_table1_summaries Remove file, permission denied. Sleeping 0.1 s and trying again.")
+      Sys.sleep(0.1)
+      file.remove(fn_in)
+    }
   } # i_var
 
   # save one file with all of the variables
